refactor scheduler main
