+++ ballista/scheduler/src/lib.rs
@@ -22,6 +22,7 @@ pub mod config;
 pub mod display;
 pub mod metrics;
 pub mod planner;
+pub mod scheduler_process;
 pub mod scheduler_server;
 #[cfg(feature = "sled")]
 pub mod standalone;
